Update resize watcher
Maintain display aspect ratio when resizing.

This is better but we can still can cut off the display with an
improper client window size.

Would be nice if we could figure out the largest aspect ratio for the
display that would fit the new client dimensions and resize the
display to that.