Modify resize watcher
Set the window size to the display size.

This maintains the aspect ratio better without cutting off the display
based on the final window client dimensions...

But it still has issues... lots of flickering because we're setting it
in a watch handler.

And for some reason just dragging the window around seems to count as
a resize event and uses intermittent values from somewhere... what's
up, SDL? Is it you, XServer?